requirements.txt
Flask
nltk
spacy
python-docx
pdfplumber
scikit-learn
numpy
en-core-web-sm==3.8.0
gunicorn
